ref: move pages inside Dashboard folder outside to be standalone
